Fix parsing of offset paramter

+++ src/boros_impl/op/read.h
@@ -9,7 +9,7 @@ typedef struct {
     Operation base;
     PyObject *buf;
     unsigned int nbytes;
-    off_t offset;
+    unsigned long long offset;
 } ReadOperation;
 
 PyObject *read_operation_create(PyObject *mod, PyObject *const *args, Py_ssize_t nargsf);
